Hardware settings bug fix
1. Fixed a bug where frames would be produced and consumed on the same
thread regardless of setting "singleThread": false.
This fix makes the marker overlay flicker possibly because it’s being
updated and read at the same time, a quick fix to this was to lock
frameLock while writing the overlay back into the image to display,
however, MarkerCamera.mm really needs a clean up.

2. Fixed the hardware settings update URL. The file was moved when the
project was split into app and core but the updateURL field was not
changed. Added a duplicate file in the old location to transition to
new location.

3. Changed the wait loop in MarkerCamera.mm to a semaphore.