Combined markers (part 2)
Better validation, adding the resetHistoryOnOpen property to markers
and default experience.
Also found some odd behaviour: if you have a marker with an invalid
action (either nil or set from an old version of the app or other
source that does not add a protocol to the start of the action), when
you read and open the marker it just turns the camera off.