Marker detection update
Changes:

- Empty regions or nested regions were ignored. Fixed by adding an else
clause “if (region!=nil) {…} else {return nil;}” in createMarkerForNode.

- It was possible to enter empty regions (0.0.1.2.3) but not recognise
them (would read as just 1.2.3). Fixed by adding a maxEmptyRegions
value calculated from the codes entered in the experience, returning
empty regions from createRegionForNode and checking this in the
validation method. Empty regions are needed for compatibility with
other topological marker systems.

- As ignoring empty regions might be desired (ignoring nested regions
is probably not) added an ignoreEmptyRegions setting that is true when
there are no codes with empty regions specified by the experience.

Note: Do not push this to the AppStore until embedded checksum is
implemented as it’s needed for the cover of Art.CHI and the removal of
ignoring nested regions means it will not work!