Added Settings.bundle
Added Settings.bundle so that features can be turned on/off, like the
Android version, from the Settings app.

Note: There is a bug with Apple’s Settings app that causes an apps
settings to not show up if the app is (re-)installed while Settings is
open, force close Settings to fix.