UI tweak for experience edit
Added a toolbar to the keyboard for editing experience name/description.
This is because the view controller’s toolbar was not appearing with
the keyboard (there was just a white bar where it should be, it’s still
there with this tweak as hiding the view controller’s toolbar causes
weird behaviour) and if you didn’t know you could swipe right to left
for the next page it was tricky to get the keyboard to disappear so you
could tap the next button on the view controller’s toolbar.