Screenshot function
Added ability to save an image of every step of detection for the next
frame (saves the overlay separately as a transparent PNG).
Triggered by new camera icon under the 3 dots menu.

Added this as I’ve needed this function for figures in a couple papers
and the operating system screenshot function creates poor quality
images.

Camera button should be hidden when used as a library. The function to
save images is in the main Artcodes project not ArtcodesScanner so apps
using the library should not need NSPhotoLibraryUsageDescription in
Info.plist