Partial update to Swift 3
The artcodes-scanner project complies and I changed the CocoaPods to
Swift 3 versions (‘pod update’ may not be enough, you may need to
delete cache files first).
Unfortunately the automated conversion process does not do much, so the
Artcodes app project still needs a lot of work and does not compile. :(